Updated class to work with staging
The paths were screwed up on the staging server because the attachment
path had the /releases/12345/ in the path whereas the home directory
used the symlink. This made narrowing the path to just the uploads was
impossible.

Now I use the URL, which has the same structure on both local and
staging.